feat: mockup session info card
