+++ components/ui/card.tsx
@@ -26,7 +26,7 @@ function CardTitle({ className, ...props }: ComponentProps<'div'>) {
   return (
     <div
       data-slot="card-title"
-      className={cn('text-lg leading-none font-medium', className)}
+      className={cn('text-xl leading-none font-medium', className)}
       {...props}
     />
   )
