+++ entrypoints/dashboard/components/info-cards/index.tsx
@@ -1,0 +1,11 @@
+import { InfoCardSession } from './info-card-session'
+
+function InfoCards() {
+  return (
+    <div className="flex w-1/3 flex-col gap-4">
+      <InfoCardSession />
+    </div>
+  )
+}
+
+export { InfoCards }
